make end effector a new type
