+++ src/lerobot/robots/widowX_250S/__init__.py
@@ -1,2 +1,3 @@
-from .widowX_250S_config import WidowX250SConfig
+from .widowX_250S_config import WidowX250SConfig, WidowX250SEndEffectorConfig
 from .widowX_250S import WidowX250S
+from .widowX_250S_end_effector import WidowX250SEndEffector
