+++ src/lerobot/robots/widowX_250S/widowX_250S.py
@@ -3,7 +3,6 @@ import time
 from functools import cached_property
 from typing import Any
 
-import numpy as np
 from lerobot.cameras.utils import make_cameras_from_configs
 from lerobot.constants import OBS_STATE
 from lerobot.errors import DeviceAlreadyConnectedError, DeviceNotConnectedError
@@ -12,13 +11,10 @@ from lerobot.motors.dynamixel import (
     DynamixelMotorsBus,
     OperatingMode,
 )
-from numpy import ndarray
-from rerun import disconnect
 
 from ..robot import Robot
 from ..utils import ensure_safe_goal_position
 from .widowX_250S_config import WidowX250SConfig
-from ...model.kinematics import RobotKinematics
 
 logger = logging.getLogger(__name__)
 
@@ -29,7 +25,7 @@ class WidowX250S(Robot):
     """
 
     config_class = WidowX250SConfig
-    name = "widowX250S"
+    name = "widowx_250s"
 
     def __init__(self, config: WidowX250SConfig):
         super().__init__(config)
@@ -53,44 +49,15 @@ class WidowX250S(Robot):
             calibration=self.calibration,
         )
 
-        # Abusing the metrics name as we use degrees and not radians for angles
-        # Obtained from https://docs.trossenrobotics.com/interbotix_xsarms_docs/specifications/wx250s.html
-        self.joint_limits_in_metrics = {
-            "waist": (-180, 180),
-            "shoulder": (-108, 114),
-            "shoulder_shadow": (-108, 114),
-            "elbow": (-123, 92),
-            "elbow_shadow": (-123, 92),
-            "forearm_roll": (-180, 180),
-            "wrist_angle": (-100, 123),
-            "wrist_rotate": (-180, 180),
-            "gripper": (30, 74),
-        }
-        self.joint_names_for_kinematics = ["waist", "shoulder", "elbow", "forearm_roll", "wrist_angle", "wrist_rotate", "gripper"]
+        self.motor_names_for_observations = ["waist", "shoulder", "elbow", "forearm_roll", "wrist_angle", "wrist_rotate", "gripper"]
 
         # --- Cameras --------------------------------------------------------
         self.cameras = make_cameras_from_configs(config.cameras)
-        # Initialize the kinematics module for the widowX250S robot
-        if self.config.urdf_path is None:
-            raise ValueError(
-                "urdf_path must be provided in the configuration for end-effector control. "
-            )
-
-        self.kinematics = RobotKinematics(
-            urdf_path=self.config.urdf_path,
-            target_frame_name=self.config.target_frame_name,
-        )
-
-        # Store the bounds for end-effector position
-        self.end_effector_bounds = self.config.end_effector_bounds
-
-        self.current_ee_pos_metrics = None
-        self.current_joint_pos_metrics = None
 
     # --------------------------- Features -----------------------------------
     @property
     def _motors_ft(self) -> dict[str, type]:
-        return {f"{motor}.pos": float for motor in self.bus.motors}
+        return {f"{motor}.pos": float for motor in self.motor_names_for_observations}
 
     @property
     def _cameras_ft(self) -> dict[str, tuple]:
@@ -216,6 +183,10 @@ class WidowX250S(Robot):
         # Read arm positions (ticks)
         start = time.perf_counter()
         obs_state = self.bus.sync_read("Present_Position")
+        # Pop shadow joints
+        obs_state.pop("shoulder_shadow", None)
+        obs_state.pop("elbow_shadow", None)
+
         obs_dict[OBS_STATE] = obs_state
         obs_dict.update({f"{motor}.pos": val for motor, val in obs_state.items()})
         dt_ms = (time.perf_counter() - start) * 1e3
@@ -256,99 +227,19 @@ class WidowX250S(Robot):
             goal_present_pos = {key: (g_pos, present_pos[key]) for key, g_pos in goal_pos.items()}
             goal_pos = ensure_safe_goal_position(goal_present_pos, self.config.max_relative_target)
 
-        # Send goal positions
-        self.bus.sync_write("Goal_Position", goal_pos)
-        return {f"{motor}.pos": val for motor, val in goal_pos.items()}
-
-    def send_action_deltas(self, action: dict[str, float]) -> dict[str, float]:
-        """Command action as deltas to the end-effector position.
-
-        The relative action magnitude may be clipped depending on the configuration
-        parameter `max_relative_target`. In this case, the action sent differs from the
-        original action. Thus, this function always returns the action actually sent.
-
-        Args:
-            action (dict[str, float]): The goal positions for the motors, keyed by
-                "<motor>.pos".
-        Returns:
-            dict[str, float]: The action sent to the motors, potentially clipped,
-                with keys "<motor>.pos".
-        """
-        if not self.is_connected:
-            raise DeviceNotConnectedError(f"{self} is not connected.")
-
-        import numpy as np
-        # Convert action to numpy array if not already
-        if isinstance(action, dict):
-            if all(k in action for k in ["delta_x", "delta_y", "delta_z"]):
-                delta_ee = np.array(
-                    [
-                        action["delta_x"] * self.config.end_effector_step_sizes["x"],
-                        action["delta_y"] * self.config.end_effector_step_sizes["y"],
-                        action["delta_z"] * self.config.end_effector_step_sizes["z"],
-                        ],
-                    dtype=np.float32,
-                )
-                if "gripper" not in action:
-                    action["gripper"] = [1.0]
-                action = np.append(delta_ee, action["gripper"])
-            else:
-                logger.warning(
-                    f"Expected action keys 'delta_x', 'delta_y', 'delta_z', got {list(action.keys())}"
-                )
-                action = np.zeros(4, dtype=np.float32)
-
-        if self.current_joint_pos_metrics is None:
-            current_joint_pos_dict = self.bus.sync_read("Present_Position", normalize=True)
-            current_joint_pos_metrics = self.normalized_joint_pos_to_metrics(current_joint_pos_dict)
-            self.current_joint_pos_metrics = np.array([current_joint_pos_metrics[name] for name in self.joint_names_for_kinematics])
-
-        # Calculate current end-effector position using forward kinematics
-        if self.current_ee_pos_metrics is None:
-            self.current_ee_pos_metrics = self.kinematics.forward_kinematics(self.current_joint_pos_metrics)
-
-        # Set desired end-effector position by adding delta
-        desired_ee_pos_metrics = np.eye(4)
-        desired_ee_pos_metrics[:3, :3] = self.current_ee_pos_metrics[:3, :3]  # Keep orientation
-
-        # Add delta to position and clip to bounds
-        desired_ee_pos_metrics[:3, 3] = self.current_ee_pos_metrics[:3, 3] + action[:3]
-        # if self.end_effector_bounds is not None:
-        desired_ee_pos_metrics[:3, 3] = np.clip(
-            desired_ee_pos_metrics[:3, 3],
-            self.end_effector_bounds["min"],
-            self.end_effector_bounds["max"],
-        )
-
-        # Compute inverse kinematics to get joint positions
-        target_joint_pos_metrics = self.kinematics.inverse_kinematics(
-            self.current_joint_pos_metrics, desired_ee_pos_metrics
-        )
-
-        # 0 is close, 1 is stay, 2 is same
-        # For now set to min and max position
-        if action[-1] < 1:
-            target_joint_pos_metrics[-1] = self.config.closed_gripper_pos
-        elif action[-1] > 1:
-            target_joint_pos_metrics[-1] = self.config.open_gripper_pos
-        else:
-            target_joint_pos_metrics[-1]  = self.current_joint_pos_metrics[-1]
-
-
-        self.current_ee_pos_metrics = desired_ee_pos_metrics.copy()
-        self.current_joint_pos_metrics = target_joint_pos_metrics.copy()
-
-        joint_metrics_dict = {key: target_joint_pos_metrics[i] for i, key in enumerate(self.joint_names_for_kinematics)}
-        # Add shadow joints and gripper
-        joint_metrics_dict["shoulder_shadow"] = joint_metrics_dict["shoulder"]
-        joint_metrics_dict["elbow_shadow"] = joint_metrics_dict["elbow"]
-
-        goal_pos = self.metrics_joint_pos_to_normalized(joint_metrics_dict)
+        # Add shadow joints
+        if "shoulder" in goal_pos:
+            goal_pos["shoulder_shadow"] = goal_pos["shoulder"]
+        if "elbow" in goal_pos:
+            goal_pos["elbow_shadow"] = goal_pos["elbow"]
 
         # Send goal positions
         self.bus.sync_write("Goal_Position", goal_pos)
+        # Pop shadow joints
+        goal_pos.pop("shoulder_shadow", None)
+        goal_pos.pop("elbow_shadow", None)
+
         return {f"{motor}.pos": val for motor, val in goal_pos.items()}
-        return {}
 
     def disconnect(self):
         if not self.is_connected:
@@ -376,48 +267,6 @@ class WidowX250S(Robot):
         for cam in self.cameras.values():
             cam.disconnect()
         logger.info(f"{self} disconnected.")
-
-    def normalized_joint_pos_to_metrics(self, norm_pos: dict[str, float]) -> dict[str, float]:
-        """Convert a normalized joint pose to degrees and mm.
-
-        Args:
-            norm_pos (dict[str, float]): Normalized joint positions, keyed by motor name.
-
-        Returns:
-            dict[str, float]: Joint positions in degrees, keyed by motor name.
-        """
-        # Unnormalize joint positions and map to degrees
-        ids_values = self.bus._get_ids_values_dict(norm_pos)
-        ids_values = self.bus._unnormalize(ids_values)
-        motor_names_to_metrics = {self.bus._id_to_name(id_): value for id_, value in ids_values.items()}
-        # Use the calibrations to convert ticks to degrees
-        for motor_name, ticks in motor_names_to_metrics.items():
-            min_deg, max_deg = self.joint_limits_in_metrics[motor_name]
-            min_tick = self.bus.calibration[motor_name].range_min
-            max_tick = self.bus.calibration[motor_name].range_max
-            motor_names_to_metrics[motor_name] = min_deg + (ticks - min_tick) * (max_deg - min_deg) / (max_tick - min_tick)
-        return motor_names_to_metrics
-
-    def metrics_joint_pos_to_normalized(self, metrics_pos: dict[str, float]) -> dict[str, float]:
-        """Convert a joint pose in degrees to normalized values.
-
-        Args:
-            metrics_pos (dict[str, float]): Joint positions in degrees, keyed by motor name.
-
-        Returns:
-            dict[str, float]: Normalized joint positions, keyed by motor name.
-        """
-        # Use the calibrations to convert degrees to ticks
-        motor_names_to_ticks = {}
-        for motor_name, deg in metrics_pos.items():
-            min_deg, max_deg = self.joint_limits_in_metrics[motor_name]
-            min_tick = self.bus.calibration[motor_name].range_min
-            max_tick = self.bus.calibration[motor_name].range_max
-            motor_names_to_ticks[motor_name] = min_tick + (deg - min_deg) * (max_tick - min_tick) / (max_deg - min_deg)
-        ids_values = self.bus._get_ids_values_dict(motor_names_to_ticks)
-        ids_values = self.bus._normalize(ids_values)
-        norm_pos = {self.bus._id_to_name(id_): value for id_, value in ids_values.items()}
-        return norm_pos
 
     def send_to_home(self) -> None:
         """Send the arm to a predefined home position."""
