+++ src/lerobot/teleoperate.py
@@ -113,7 +113,7 @@ def teleop_loop(
             observation = robot.get_observation()
             log_rerun_data(observation, action)
 
-        robot.send_action_deltas(action)
+        robot.send_action(action)
         dt_s = time.perf_counter() - loop_start
         busy_wait(1 / fps - dt_s)
